fix warnings and improve xmldocs
