+++ src/ImageSharp/Advanced/AdvancedImageExtensions.cs
@@ -18,7 +18,6 @@ namespace SixLabors.ImageSharp.Advanced
         /// <summary>
         /// Gets the configuration for the image.
         /// </summary>
-        /// <typeparam name="TPixel">The Pixel format.</typeparam>
         /// <param name="source">The source image.</param>
         /// <returns>Returns the configuration.</returns>
         public static Configuration GetConfiguration(this Image source)
