+++ src/ImageSharp/Formats/Bmp/BmpDecoder.cs
@@ -30,6 +30,7 @@ namespace SixLabors.ImageSharp.Formats.Bmp
             return new BmpDecoderCore(configuration, this).Decode<TPixel>(stream);
         }
 
+        /// <inheritdoc />
         public Image Decode(Configuration configuration, Stream stream) => this.Decode<Rgba32>(configuration, stream);
 
         /// <inheritdoc/>
