+++ src/ImageSharp/Formats/Gif/GifDecoder.cs
@@ -44,7 +44,8 @@ namespace SixLabors.ImageSharp.Formats.Gif
             var decoder = new GifDecoderCore(configuration, this);
             return decoder.Identify(stream);
         }
-        
+
+        /// <inheritdoc />
         public Image Decode(Configuration configuration, Stream stream) => this.Decode<Rgba32>(configuration, stream);
     }
 }
