+++ src/ImageSharp/Formats/IImageDecoder.cs
@@ -12,15 +12,22 @@ namespace SixLabors.ImageSharp.Formats
     public interface IImageDecoder
     {
         /// <summary>
-        /// Decodes the image from the specified stream to the <see cref="ImageFrame{TPixel}"/>.
+        /// Decodes the image from the specified stream to an <see cref="Image{TPixel}"/> of a specific pixel type.
         /// </summary>
         /// <typeparam name="TPixel">The pixel format.</typeparam>
         /// <param name="configuration">The configuration for the image.</param>
         /// <param name="stream">The <see cref="Stream"/> containing image data.</param>
-        /// <returns>The decoded image</returns>
+        /// <returns>The decoded image of a given pixel type.</returns>
         Image<TPixel> Decode<TPixel>(Configuration configuration, Stream stream)
             where TPixel : struct, IPixel<TPixel>;
 
+        /// <summary>
+        /// Decodes the image from the specified stream to an <see cref="Image"/>.
+        /// The decoder is free to choose the pixel type.
+        /// </summary>
+        /// <param name="configuration">The configuration for the image.</param>
+        /// <param name="stream">The <see cref="Stream"/> containing image data.</param>
+        /// <returns>The decoded image of a pixel type chosen by the decoder.</returns>
         Image Decode(Configuration configuration, Stream stream);
     }
 }
