+++ src/ImageSharp/Processing/IImageProcessingContext{TPixel}.cs
@@ -3,32 +3,12 @@
 
 using SixLabors.ImageSharp.PixelFormats;
 using SixLabors.ImageSharp.Processing.Processors;
-using SixLabors.Memory;
 using SixLabors.Primitives;
 
 namespace SixLabors.ImageSharp.Processing
 {
-    public interface IImageProcessingContext
-    {
-        /// <summary>
-        /// Gets a reference to the <see cref="MemoryAllocator" /> used to allocate buffers
-        /// for this context.
-        /// </summary>
-        MemoryAllocator MemoryAllocator { get; }
-
-        /// <summary>
-        /// Gets the image dimensions at the current point in the processing pipeline.
-        /// </summary>
-        /// <returns>The <see cref="Rectangle"/></returns>
-        Size GetCurrentSize();
-
-        IImageProcessingContext ApplyProcessor(IImageProcessor processor, Rectangle rectangle);
-
-        IImageProcessingContext ApplyProcessor(IImageProcessor processor);
-    }
-
     /// <summary>
-    /// An interface to queue up image operations to apply to an image.
+    /// A pixel-specific interface to queue up image operations to apply to an image.
     /// </summary>
     /// <typeparam name="TPixel">The pixel format</typeparam>
     public interface IImageProcessingContext<TPixel> : IImageProcessingContext
@@ -37,8 +17,8 @@ namespace SixLabors.ImageSharp.Processing
         /// <summary>
         /// Adds the processor to the current set of image operations to be applied.
         /// </summary>
-        /// <param name="processor">The processor to apply</param>
-        /// <param name="rectangle">The area to apply it to</param>
+        /// <param name="processor">The processor to apply.</param>
+        /// <param name="rectangle">The area to apply it to.</param>
         /// <returns>The current operations class to allow chaining of operations.</returns>
         IImageProcessingContext<TPixel> ApplyProcessor(IImageProcessor<TPixel> processor, Rectangle rectangle);
