+++ src/ImageSharp/Processing/ProcessingExtensions.cs
@@ -13,43 +13,6 @@ namespace SixLabors.ImageSharp.Processing
     /// </summary>
     public static class ProcessingExtensions
     {
-        class ProcessingVisitor : IImageVisitor
-        {
-            private readonly Action<IImageProcessingContext> operation;
-
-            private readonly bool mutate;
-            
-            public Image ResultImage { get; private set; }
-
-            public ProcessingVisitor(Action<IImageProcessingContext> operation, bool mutate)
-            {
-                this.operation = operation;
-                this.mutate = mutate;
-            }
-
-            public void Visit<TPixel>(Image<TPixel> image)
-                where TPixel : struct, IPixel<TPixel>
-            {
-                IInternalImageProcessingContext<TPixel> operationsRunner = image.GetConfiguration()
-                    .ImageOperationsProvider.CreateImageProcessingContext(image, this.mutate);
-                this.operation(operationsRunner);
-                this.ResultImage = operationsRunner.Apply();
-            }
-        }
-        
-        public static void Mutate(this Image source, Action<IImageProcessingContext> operation)
-        {
-            ProcessingVisitor visitor = new ProcessingVisitor(operation, true);
-            source.AcceptVisitor(visitor);
-        }
-
-        public static Image Clone(this Image source, Action<IImageProcessingContext> operation)
-        {
-            ProcessingVisitor visitor = new ProcessingVisitor(operation, false);
-            source.AcceptVisitor(visitor);
-            return visitor.ResultImage;
-        }
-        
         /// <summary>
         /// Applies the given operation to the mutable image.
         /// Useful when we need to extract information like Width/Height to parametrize the next operation working on the <see cref="IImageProcessingContext{TPixel}"/> chain.
@@ -61,6 +24,17 @@ namespace SixLabors.ImageSharp.Processing
         /// <returns>The <see cref="IImageProcessingContext{TPixel}"/> to allow chaining of operations.</returns>
         public static IImageProcessingContext<TPixel> Apply<TPixel>(this IImageProcessingContext<TPixel> source, Action<Image<TPixel>> operation)
             where TPixel : struct, IPixel<TPixel> => source.ApplyProcessor(new DelegateProcessor<TPixel>(operation));
+
+        /// <summary>
+        /// Mutates the source image by applying the image operation to it.
+        /// </summary>
+        /// <param name="source">The image to mutate.</param>
+        /// <param name="operation">The operation to perform on the source.</param>
+        public static void Mutate(this Image source, Action<IImageProcessingContext> operation)
+        {
+            ProcessingVisitor visitor = new ProcessingVisitor(operation, true);
+            source.AcceptVisitor(visitor);
+        }
 
         /// <summary>
         /// Mutates the source image by applying the image operation to it.
@@ -94,6 +68,19 @@ namespace SixLabors.ImageSharp.Processing
             IInternalImageProcessingContext<TPixel> operationsRunner = source.GetConfiguration().ImageOperationsProvider.CreateImageProcessingContext(source, true);
             operationsRunner.ApplyProcessors(operations);
             operationsRunner.Apply();
+        }
+
+        /// <summary>
+        /// Creates a deep clone of the current image. The clone is then mutated by the given operation.
+        /// </summary>
+        /// <param name="source">The image to clone.</param>
+        /// <param name="operation">The operation to perform on the clone.</param>
+        /// <returns>The new <see cref="SixLabors.ImageSharp.Image"/>.</returns>
+        public static Image Clone(this Image source, Action<IImageProcessingContext> operation)
+        {
+            ProcessingVisitor visitor = new ProcessingVisitor(operation, false);
+            source.AcceptVisitor(visitor);
+            return visitor.ResultImage;
         }
 
         /// <summary>
@@ -148,6 +135,30 @@ namespace SixLabors.ImageSharp.Processing
             }
 
             return source;
+        }
+
+        private class ProcessingVisitor : IImageVisitor
+        {
+            private readonly Action<IImageProcessingContext> operation;
+
+            private readonly bool mutate;
+
+            public ProcessingVisitor(Action<IImageProcessingContext> operation, bool mutate)
+            {
+                this.operation = operation;
+                this.mutate = mutate;
+            }
+
+            public Image ResultImage { get; private set; }
+
+            public void Visit<TPixel>(Image<TPixel> image)
+                where TPixel : struct, IPixel<TPixel>
+            {
+                IInternalImageProcessingContext<TPixel> operationsRunner = image.GetConfiguration()
+                    .ImageOperationsProvider.CreateImageProcessingContext(image, this.mutate);
+                this.operation(operationsRunner);
+                this.ResultImage = operationsRunner.Apply();
+            }
         }
     }
 }
