+++ src/ImageSharp/Processing/Processors/Filters/LomographProcessorImplementation.cs
@@ -7,20 +7,23 @@ using SixLabors.Primitives;
 
 namespace SixLabors.ImageSharp.Processing.Processors.Filters
 {
+    /// <summary>
+    /// Converts the colors of the image recreating an old Lomograph effect.
+    /// </summary>
     internal class LomographProcessorImplementation<TPixel> : FilterProcessorImplementation<TPixel>
         where TPixel : struct, IPixel<TPixel>
     {
         private static readonly TPixel VeryDarkGreen = ColorBuilder<TPixel>.FromRGBA(0, 10, 0, 255);
 
+        public LomographProcessorImplementation(LomographProcessor definition)
+            : base(definition)
+        {
+        }
+
         /// <inheritdoc/>
         protected override void AfterFrameApply(ImageFrame<TPixel> source, Rectangle sourceRectangle, Configuration configuration)
         {
             new VignetteProcessor<TPixel>(VeryDarkGreen).Apply(source, sourceRectangle, configuration);
-        }
-
-        public LomographProcessorImplementation(FilterProcessor definition)
-            : base(definition)
-        {
         }
     }
 }
