+++ src/ImageSharp/Processing/Processors/Filters/OpacityProcessor.cs
@@ -8,11 +8,10 @@ namespace SixLabors.ImageSharp.Processing.Processors.Filters
     /// <summary>
     /// Applies an opacity filter matrix using the given amount.
     /// </summary>
-    /// <typeparam name="TPixel">The pixel format.</typeparam>
     internal class OpacityProcessor : FilterProcessor
     {
         /// <summary>
-        /// Initializes a new instance of the <see cref="OpacityProcessor{TPixel}"/> class.
+        /// Initializes a new instance of the <see cref="OpacityProcessor"/> class.
         /// </summary>
         /// <param name="amount">The proportion of the conversion. Must be between 0 and 1.</param>
         public OpacityProcessor(float amount)
