+++ src/ImageSharp/Processing/Processors/Filters/PolaroidProcessor.cs
@@ -1,10 +1,6 @@
 ﻿// Copyright (c) Six Labors and contributors.
 // Licensed under the Apache License, Version 2.0.
 
-using SixLabors.ImageSharp.PixelFormats;
-using SixLabors.ImageSharp.Processing.Processors.Overlays;
-using SixLabors.Primitives;
-
 namespace SixLabors.ImageSharp.Processing.Processors.Filters
 {
     /// <summary>
@@ -13,33 +9,15 @@ namespace SixLabors.ImageSharp.Processing.Processors.Filters
     public class PolaroidProcessor : FilterProcessor
     {
         /// <summary>
-        /// Initializes a new instance of the <see cref="PolaroidProcessor{TPixel}" /> class.
+        /// Initializes a new instance of the <see cref="PolaroidProcessor" /> class.
         /// </summary>
         public PolaroidProcessor()
             : base(KnownFilterMatrices.PolaroidFilter)
         {
         }
 
+        /// <inheritdoc />
         public override IImageProcessor<TPixel> CreatePixelSpecificProcessor<TPixel>() =>
             new PolaroidProcessorImplementation<TPixel>(this);
-    }
-
-    internal class PolaroidProcessorImplementation<TPixel> : FilterProcessorImplementation<TPixel>
-        where TPixel : struct, IPixel<TPixel>
-    {
-        private static readonly TPixel VeryDarkOrange = ColorBuilder<TPixel>.FromRGB(102, 34, 0);
-        private static readonly TPixel LightOrange = ColorBuilder<TPixel>.FromRGBA(255, 153, 102, 128);
-
-        /// <inheritdoc/>
-        protected override void AfterFrameApply(ImageFrame<TPixel> source, Rectangle sourceRectangle, Configuration configuration)
-        {
-            new VignetteProcessor<TPixel>(VeryDarkOrange).Apply(source, sourceRectangle, configuration);
-            new GlowProcessor<TPixel>(LightOrange, source.Width / 4F).Apply(source, sourceRectangle, configuration);
-        }
-
-        public PolaroidProcessorImplementation(FilterProcessor definition)
-            : base(definition)
-        {
-        }
     }
 }
